+++ R/fit_growth.R
@@ -15,7 +15,8 @@
 #' fit_growth(df=mydata, Time, OD600, type = "logistic")}
 #'
 fit_growth <- function(df, time, data, type = "parametric", ...) {
-    fit_growth_(df, time_col=lazy(time), data_col=lazy(data), type = type, ...)
+    fit_growth_(df, time_col = lazy(time), data_col = lazy(data),
+                type = type, ...)
 }
 
 
@@ -23,14 +24,21 @@ fit_growth <- function(df, time, data, type = "parametric", ...) {
 #' @rdname fit_growth
 #' @importFrom grofit grofit.control
 fit_growth_ <- function(df, time_col, data_col, type = "parametric", ...) {
+    
+    if (!type %in% c("parametric", "logistic", "gompertz", "gompertz.exp",
+                    "richards", "spline")) {
+        stop(sprintf("Unsupported model type '%s'", type))
+    }
+    
     if (identical(type, "spline")) {
-        fit_growth_spline_(df, time_col=time_col, data_col=data_col, ...)
+        fit_growth_spline_(df, time_col = time_col, data_col = data_col, ...)
     }
     else if (identical(type, "parametric")) {
-        fit_growth_parametric_(df, time_col=time_col, data_col=data_col, ...)
+        fit_growth_parametric_(df, time_col = time_col, data_col = data_col,
+                               ...)
     }
     else {
-        fit_growth_parametric_(df, time_col=time_col, data_col=data_col,
-                               control=grofit.control(model.type=type), ...)
+        fit_growth_parametric_(df, time_col = time_col, data_col = data_col,
+                               control = grofit.control(model.type=type), ...)
     }
 }