+++ R/fit_growth_gompertz.R
@@ -2,7 +2,7 @@
 #'
 #' @inheritParams fit_growth_parametric
 #'
-#' @importFrom lazyeval lazy
+#' @importFrom grofit grofit.control
 #' @seealso \code{\link{fit_growth_parametric}}
 #' @seealso \code{\link{gompertz}}
 #' @seealso \url{https://en.wikipedia.org/wiki/Gompertz_function}
@@ -13,8 +13,8 @@
 #' # Fit the data given in columns Time and OD600
 #' fit_growth_gompertz(mydata, Time, OD600)}
 fit_growth_gompertz <- function(df, time, data, ...) {
-    fit_growth_parametric_(df, time_col=lazy(time), data_col=lazy(data),
-                           control=grofit.control(model.type="gompertz"), ...)
+    fit_growth_parametric(df, time = time, data = data,
+                          control = grofit.control(model.type="gompertz"), ...)
 }
 
 #' @export
@@ -27,6 +27,6 @@ fit_growth_gompertz <- function(df, time, data, ...) {
 #' fit_growth_gompertz_(df=mydata, time_col='Time', data_col='OD600')}
 #'
 fit_growth_gompertz_ <- function(df, time_col, data_col, ...) {
-    fit_growth_parametric_(df, time_col=time_col, data_col=data_col,
-                           control=grofit.control(model.type="gompertz"), ...)
+    fit_growth_parametric_(df = df, time_col = time_col, data_col = data_col,
+                           control = grofit.control(model.type="gompertz"), ...)
 }