+++ R/fit_growth_logistic.R
@@ -2,7 +2,7 @@
 #'
 #' @inheritParams fit_growth_parametric
 #'
-#' @importFrom lazyeval lazy
+#' @importFrom grofit grofit.control
 #' @seealso \code{\link{fit_growth_parametric}}
 #' @seealso \code{\link{logistic}}
 #' @seealso \url{https://en.wikipedia.org/wiki/Logistic_function}
@@ -14,8 +14,8 @@
 #' fit_growth_logistic(mydata, Time, OD600)}
 #'
 fit_growth_logistic <- function(df, time, data, ...) {
-    fit_growth_parametric_(df, time_col=lazy(time), data_col=lazy(data),
-                           control=grofit.control(model.type="logistic"), ...)
+    fit_growth_parametric(df, time = time, data = data,
+                           control = grofit.control(model.type="logistic"), ...)
 }
 
 
@@ -29,6 +29,6 @@ fit_growth_logistic <- function(df, time, data, ...) {
 #' fit_growth_logistic_(df=mydata, time_col='Time', data_col='OD600')}
 #'
 fit_growth_logistic_ <- function(df, time_col, data_col, ...) {
-    fit_growth_parametric_(df, time_col=time_col, data_col=data_col,
-                           control=grofit.control(model.type="logistic"), ...)
+    fit_growth_parametric_(df = df, time_col = time_col, data_col = data_col,
+                           control = grofit.control(model.type="logistic"), ...)
 }